Made it create a file if it does't exist
Made the script check if the file doesn't exist so that Subl wouldn't receive the ~20 word string error from wslpath and create those files. Also added -w (wait) argument to subl.exe so that it waits for Sublime Text to close before writing to the file. Additonally, I got rid of the cwd=$(pwd); because it is not used. Lastly, if no arguments are given, it assumes that you just want to open Sublime Text.